clustering/image-compressor-kmeans/requirements.txt
flask
numpy
pandas
scikit-learn
joblib
pillow
matplotlib
